Refactor multiclass utilities into a separate module
Moved multiclass utility functions from `__init__.py` to a new module, `tmp.py`, to improve code organization. This change isolates utility functions, making the codebase more modular and maintainable.
